Also grant expand user access.
With out this change, if access was provisioned via a user entitlement will be moved to a group assignment, because that's how the access is actually established. So no matter how you got the access you will now both have a user and a group entitlement for that access to an app. Only the user one is requestable. I am not aware of a workaround